Merge in 32-bit-st.

+++ src/mod_minimizer.cpp
@@ -6,7 +6,7 @@ namespace digest{
 
 		if(get_minimized_h() == digest::MinimizedHashType::CANON) {
 			do {
-                if(chash % mod == congruence){
+                if((uint32_t)chash % mod == congruence){
                     vec.push_back(get_pos());
                 }
 			} while(roll_one() && vec.size() < amount);
@@ -15,7 +15,7 @@ namespace digest{
 
         if(get_minimized_h() == digest::MinimizedHashType::FORWARD) {
 			do {
-                if(fhash % mod == congruence){
+                if((uint32_t)fhash % mod == congruence){
                     vec.push_back(get_pos());
                 }
 			} while(roll_one() && vec.size() < amount);
@@ -24,7 +24,7 @@ namespace digest{
 
 		// reverse
 		do {
-			if(rhash % mod == congruence){
+			if((uint32_t)rhash % mod == congruence){
 				vec.push_back(get_pos());
 			}
         } while(roll_one() && vec.size() < amount);
